added authentication check for DELETE method / exercise 4.21

+++ part4/server/controllers/blogs.js
@@ -39,7 +39,24 @@ blogsRouter.post("/", async (request, response) => {
 })
 
 blogsRouter.delete("/:id", async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if(!request.token || !decodedToken.id){
+        return response.status(401).json({error: "token invalid"})
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    const blog = await Blog.findById(request.params.id)
+
+    if(blog.user.toString() === user.id.toString()){
+        await Blog.findByIdAndRemove(request.params.id)
+    }else{
+        return response.status(401).json({error: "no authentication on selected item"})
+    }
+
+    
     response.status(204).end()
 })
